requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
lightning==2.5.1
lightning-utilities==0.14.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127con
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.11.0.86
packaging==24.2
pillow==11.2.1
propcache==0.3.1
pytorch-lightning==2.5.1
PyYAML==6.0.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchmetrics==1.7.1
torchvision==0.21.0+cu124
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
yarl==1.20.0
